reorders the linker sections; make the reset_handler the entry point
Before this commit the .rodata sections of the inputs were placed in the output
.text section. As the vector table is in the .rodata of the cortex-m-rt crate
this placed the vector table first in the output .text section:

``` console
$ arm-none-eabi-objdump -Cd hello
Disassembly of section .text:

08000000 <_VECTOR_TABLE>:
 8000000:       10002000        .word   0x10002000

08000004 <cortex_m_rt::RESET_HANDLER>:
 8000004:       08000405                                ....
(..)

$ arm-none-eabi-readelf -h hello | grep Entry
  Entry point address:               0x8000000
```

With this memory layout using `load` and then `step` on a GDB session made the
processor execute the vector table (PC = 0x08000000).

This commit places the .rodata sections of the inputs into the output .rodata
section, and also moves the reset_handler code to the start the output .text
section. This turns the reset_handler into the entry point of the program.

``` console
$ arm-none-eabi-objdump -Cd hello
Disassembly of section .text:

08000410 <cortex_m_rt::reset_handler>:
(..)

$ arm-none-eabi-readelf -h hello | grep Entry
  Entry point address:               0x8000410
```

With these changes the GDB commands `load` and `step` work correctly: `load`
sets the PC = reset_handler.

Also the `size` output is now more helpful:

``` console
$ arm-none-eabi-size -Ax hello
section                 size         addr
.vector_table          0x400    0x8000000
.text                  0x120    0x8000400
.rodata                  0xd    0x8000520
.bss                     0x4   0x20000000
.data                    0x4   0x20000004
```