discard the ARM.exidx.* section
It's related to stack unwinding and takes up FLASH memory. I doubt anyone is using it and we have
been discarding this section in all the released versions of this crate so this is not a breaking
change.

In the future, if need for this arises, we can add it back behind a Cargo feature, or without one if
that makes sense.