add NVIC::{mask,unmask}
these are the "static method" (methods that don't take `self`) versions of
NVIC::{enable,disable} in the same vein as the existing NVIC::{pend,unpend}

this commit also deprecates the existing NVIC::{enable,disable} methods and
notes that NVIC::enable is unsound because it should be an `unsafe` method (like
interrupt::enable and basepri::write, it can break critical sections) but it's
marked as safe. Its replacement, NVIC::unmask, has the correct unsafety setting:
it's an `unsafe` function.