Remove 'extern crate cortex_m_rt' from macros
See justification on 519d46a. Using 'extern crate cortex_m_rt'
inside of the macros limits our ability to use the macros
crate in other contexts. As long as another crate publicly
exports an enumeration of `interrupt` and an enumeration
of `exception`, the macros crate may be used in other
cortex-m-rt-like systems.