macros: Allow setting a name for singleton!() statics
Right now, the `singleton!()` macro always creates a static named `VAR`.
This is annoying when digging through objdump output or digging into
memory with a debugger because it is hard to see what singleton you're
looking at when they are all called `<...>::{{closure}}::VAR`.

To make it a bit nicer, allow setting the name of the generated static
to some meaningful identifier which can then be cross-referenced in
debugger output, for example with

	singleton!(FOO_BUFFER: [u8; 1024] = [0u8; 1024]);

There is no other side-effects to this change - the identifier is never
visible to other code because it is still contained in the closure of
the macro.