CI: reduce duration
The data_overflow test takes 3 minutes to compile on my system.
On the GitHub actions runners it takes 10-15 minutes.
I do not know the underlying reason for this.

Splitting up the arrays in data_overflow into smaller 8k chunks does not
change the test outcome, but rustc compiles it in 5 seconds (36x faster).