Optimize critical_section: use MSR instruction
To avoid branches and unnecessary code in the acquire and release
functions, simply write back the original value to the PRIMASK register
using the MSR instruction.

This changes the critical-section dependency to use `u32` as the
`RawRestoreState` type instead of `bool`.

The `register::primask` module now defines a `struct(pub u32)` instead of an
`enum`, and additionally a `write` function for use with the critical
section.