We are in a mess. This commit attempts to fix it.
The cortex-m crate in the master branch is 0.7.4 plus a bunch of fixes
that have never been released. Some of them are breaking changes and so
will never be released. We cannot 'just' release a cortex-m 0.8 with
them in because that is ecosystem shattering - the crate contains a
'links=' attribute, and a bunch of no_mangle functions, so you cannot
have two versions in the same dependency tree.

Instead, there is a 0.7.x branch. This was taken before the cortex-m-rt
crate move. It has cortex-m 0.7.7, which has been published.

This commit just deletes the cortex-m folder and replaces it with
the contents of the cortex-m crate as found on the the 0.7.x branch.

If you want to pick up the changes that have just been deleted, cherry-
pick them into a new branch and open a PR to merge them into here.
Breaking changes for cortex-m into master will no longer be accepted.
Sorry.